+++ e2e-demo/js/index.js
@@ -143,6 +143,8 @@ let usernameMenuButton = document.getElementById('username-menu-button');
 let disableAudioSend = document.getElementById('control__disable_audio_send');
 let disableVideoSend = document.getElementById('control__disable_video_send');
 let disableAudioRecv = document.getElementById('control__disable_audio_recv');
+let disableDisplayAudioRecv = document.getElementById('control__disable_display_audio_recv');
+let disableDeviceAudioRecv = document.getElementById('control__disable_device_audio_recv');
 let disableVideoRecv = document.getElementById('control__disable_video_recv');
 let disableDisplayVideoRecv = document.getElementById('control__disable_display_video_recv');
 let disableCameraDisplayVideoRecv = document.getElementById('control__disable_camera_video_recv');
@@ -633,6 +635,8 @@ window.onload = async function() {
   let isAudioSendEnabled = true;
   let isVideoSendEnabled = true;
   let isAudioRecvEnabled = true;
+  let isDisplayAudioRecvEnabled = true;
+  let isDeviceAudioRecvEnabled = true;
   let isVideoRecvEnabled = true;
   let isDisplayVideoRecvEnabled = true;
   let isCameraVideoRecvEnabled = true;
@@ -721,7 +725,7 @@ window.onload = async function() {
         audio.device_id(audioSource.value);
       }
       audio.exact_auto_gain_control(!disableAudioGainControlSwitchEl.checked);
-      constraints.audio(audio);
+      constraints.device_audio(audio);
     }
 
     if (video_select != null) {
@@ -742,6 +746,7 @@ window.onload = async function() {
           constraints.device_video(video);
           if (screenshareSwitchEl.checked) {
             constraints.display_video(new rust.DisplayVideoTrackConstraints());
+            constraints.display_audio(new rust.AudioTrackConstraints());
           }
         }
       } else {
@@ -893,21 +898,39 @@ window.onload = async function() {
             playElement.firstChild.srcObject = mediaStream;
           }
         } else {
-          playElement = memberVideoDiv.getElementsByClassName('audio')[0];
-          if (playElement === undefined) {
-            playElement = document.createElement('audio');
-            playElement.classList.add('audio', 'order-3');
-            playElement.controls = 'true';
-            playElement.autoplay = 'true';
-            if (track.media_direction() != 0) {
-              playElement.pause();
-              playElement.style.display = 'none';
+          if (track.media_source_kind() === rust.MediaSourceKind.Display) {
+            playElement = memberVideoDiv.getElementsByClassName('display-audio')[0];
+            if (playElement === undefined) {
+              playElement = document.createElement('audio');
+              playElement.classList.add('display-audio', 'order-3');
+              playElement.controls = 'true';
+              playElement.autoplay = 'true';
+              if (track.media_direction() !== 0) {
+                playElement.pause();
+                playElement.style.display = 'none';
+              }
+              memberVideoDiv.appendChild(playElement);
             }
-            memberVideoDiv.appendChild(playElement);
+            let mediaStream = new MediaStream();
+            mediaStream.addTrack(track.get_track());
+            playElement.srcObject = mediaStream;
+          } else {
+            playElement = memberVideoDiv.getElementsByClassName('device-audio')[0];
+            if (playElement === undefined) {
+              playElement = document.createElement('audio');
+              playElement.classList.add('device-audio', 'order-3');
+              playElement.controls = 'true';
+              playElement.autoplay = 'true';
+              if (track.media_direction() !== 0) {
+                playElement.pause();
+                playElement.style.display = 'none';
+              }
+              memberVideoDiv.appendChild(playElement);
+            }
+            let mediaStream = new MediaStream();
+            mediaStream.addTrack(track.get_track());
+            playElement.srcObject = mediaStream;
           }
-          let mediaStream = new MediaStream();
-          mediaStream.addTrack(track.get_track());
-          playElement.srcObject = mediaStream;
         }
 
         track.on_media_direction_changed(async (direction) => {
@@ -1035,7 +1058,8 @@ window.onload = async function() {
           }
           await room.set_local_media_settings(constraints, true, true);
           if (screenshareSwitchEl.checked) {
-            await room.enable_video(rust.MediaSourceKind.Display);
+              await room.enable_video(rust.MediaSourceKind.Display);
+              await room.enable_audio(rust.MediaSourceKind.Display);
           }
         } catch (e) {
           let name = e.kind();
@@ -1147,6 +1171,28 @@ window.onload = async function() {
         await room.enable_remote_audio();
         isAudioRecvEnabled = true;
         disableAudioRecv.textContent = 'Disable audio recv'
+      }
+    });
+    disableDisplayAudioRecv.addEventListener('click', async () => {
+      if (isDisplayAudioRecvEnabled) {
+        await room.disable_remote_audio(rust.MediaSourceKind.Display);
+        isDisplayAudioRecvEnabled = false;
+        disableDisplayAudioRecv.textContent = 'Enable display audio recv'
+      } else {
+        await room.enable_remote_audio(rust.MediaSourceKind.Display);
+        isDisplayAudioRecvEnabled = true;
+        disableDisplayAudioRecv.textContent = 'Disable display audio recv'
+      }
+    });
+    disableDeviceAudioRecv.addEventListener('click', async () => {
+      if (isDeviceAudioRecvEnabled) {
+        await room.disable_remote_audio(rust.MediaSourceKind.Device);
+        isDeviceAudioRecvEnabled = false;
+        disableDeviceAudioRecv.textContent = 'Enable device audio recv'
+      } else {
+        await room.enable_remote_audio(rust.MediaSourceKind.Device);
+        isDeviceAudioRecvEnabled = true;
+        disableDeviceAudioRecv.textContent = 'Disable device audio recv'
       }
     });
     disableVideoRecv.addEventListener('click', async () => {