+++ e2e/src/object/jason.rs
@@ -81,7 +81,7 @@ impl Object<Jason> {
 
                 let constraints = new rust.MediaStreamSettings();
                 let audio = new window.rust.AudioTrackConstraints();
-                constraints.audio(audio);
+                constraints.device_audio(audio);
                 let video = new window.rust.DeviceVideoTrackConstraints();
                 constraints.device_video(video);
                 await room