+++ src/api/dart/api/api_bridge_generated.rs
@@ -1283,7 +1283,8 @@ fn wire__crate__api__dart__api__room__RoomHandle_disable_audio_impl(
     FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec,_>(flutter_rust_bridge::for_generated::TaskInfo{ debug_name: "RoomHandle_disable_audio", port: None, mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync }, move || { 
             let message = unsafe { flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(ptr_, rust_vec_len_, data_len_) };
             let mut deserializer = flutter_rust_bridge::for_generated::SseDeserializer::new(message);
-            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);deserializer.end();
+            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);
+let api_source_kind = <Option<crate::media::track::MediaSourceKind>>::sse_decode(&mut deserializer);deserializer.end();
                 transform_result_sse::<_, ()>((move || {
                     let mut api_that_guard = None;
 let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decode_order(vec![flutter_rust_bridge::for_generated::LockableOrderInfo::new(&api_that, 0, false)]);
@@ -1294,7 +1295,7 @@ let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decod
             }
         }
         let api_that_guard = api_that_guard.unwrap();
- let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::disable_audio(&*api_that_guard))?;   Ok(output_ok)
+ let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::disable_audio(&*api_that_guard, api_source_kind))?;   Ok(output_ok)
                 })()) })
 }
 fn wire__crate__api__dart__api__room__RoomHandle_disable_remote_audio_impl(
@@ -1305,7 +1306,8 @@ fn wire__crate__api__dart__api__room__RoomHandle_disable_remote_audio_impl(
     FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec,_>(flutter_rust_bridge::for_generated::TaskInfo{ debug_name: "RoomHandle_disable_remote_audio", port: None, mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync }, move || { 
             let message = unsafe { flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(ptr_, rust_vec_len_, data_len_) };
             let mut deserializer = flutter_rust_bridge::for_generated::SseDeserializer::new(message);
-            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);deserializer.end();
+            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);
+let api_source_kind = <Option<crate::media::track::MediaSourceKind>>::sse_decode(&mut deserializer);deserializer.end();
                 transform_result_sse::<_, ()>((move || {
                     let mut api_that_guard = None;
 let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decode_order(vec![flutter_rust_bridge::for_generated::LockableOrderInfo::new(&api_that, 0, false)]);
@@ -1316,7 +1318,7 @@ let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decod
             }
         }
         let api_that_guard = api_that_guard.unwrap();
- let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::disable_remote_audio(&*api_that_guard))?;   Ok(output_ok)
+ let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::disable_remote_audio(&*api_that_guard, api_source_kind))?;   Ok(output_ok)
                 })()) })
 }
 fn wire__crate__api__dart__api__room__RoomHandle_disable_remote_video_impl(
@@ -1373,7 +1375,8 @@ fn wire__crate__api__dart__api__room__RoomHandle_enable_audio_impl(
     FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec,_>(flutter_rust_bridge::for_generated::TaskInfo{ debug_name: "RoomHandle_enable_audio", port: None, mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync }, move || { 
             let message = unsafe { flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(ptr_, rust_vec_len_, data_len_) };
             let mut deserializer = flutter_rust_bridge::for_generated::SseDeserializer::new(message);
-            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);deserializer.end();
+            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);
+let api_source_kind = <Option<crate::media::track::MediaSourceKind>>::sse_decode(&mut deserializer);deserializer.end();
                 transform_result_sse::<_, ()>((move || {
                     let mut api_that_guard = None;
 let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decode_order(vec![flutter_rust_bridge::for_generated::LockableOrderInfo::new(&api_that, 0, false)]);
@@ -1384,7 +1387,7 @@ let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decod
             }
         }
         let api_that_guard = api_that_guard.unwrap();
- let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::enable_audio(&*api_that_guard))?;   Ok(output_ok)
+ let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::enable_audio(&*api_that_guard, api_source_kind))?;   Ok(output_ok)
                 })()) })
 }
 fn wire__crate__api__dart__api__room__RoomHandle_enable_remote_audio_impl(
@@ -1395,7 +1398,8 @@ fn wire__crate__api__dart__api__room__RoomHandle_enable_remote_audio_impl(
     FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec,_>(flutter_rust_bridge::for_generated::TaskInfo{ debug_name: "RoomHandle_enable_remote_audio", port: None, mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync }, move || { 
             let message = unsafe { flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(ptr_, rust_vec_len_, data_len_) };
             let mut deserializer = flutter_rust_bridge::for_generated::SseDeserializer::new(message);
-            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);deserializer.end();
+            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);
+let api_source_kind = <Option<crate::media::track::MediaSourceKind>>::sse_decode(&mut deserializer);deserializer.end();
                 transform_result_sse::<_, ()>((move || {
                     let mut api_that_guard = None;
 let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decode_order(vec![flutter_rust_bridge::for_generated::LockableOrderInfo::new(&api_that, 0, false)]);
@@ -1406,7 +1410,7 @@ let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decod
             }
         }
         let api_that_guard = api_that_guard.unwrap();
- let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::enable_remote_audio(&*api_that_guard))?;   Ok(output_ok)
+ let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::enable_remote_audio(&*api_that_guard, api_source_kind))?;   Ok(output_ok)
                 })()) })
 }
 fn wire__crate__api__dart__api__room__RoomHandle_enable_remote_video_impl(
@@ -1486,7 +1490,8 @@ fn wire__crate__api__dart__api__room__RoomHandle_mute_audio_impl(
     FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec,_>(flutter_rust_bridge::for_generated::TaskInfo{ debug_name: "RoomHandle_mute_audio", port: None, mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync }, move || { 
             let message = unsafe { flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(ptr_, rust_vec_len_, data_len_) };
             let mut deserializer = flutter_rust_bridge::for_generated::SseDeserializer::new(message);
-            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);deserializer.end();
+            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);
+let api_source_kind = <Option<crate::media::track::MediaSourceKind>>::sse_decode(&mut deserializer);deserializer.end();
                 transform_result_sse::<_, ()>((move || {
                     let mut api_that_guard = None;
 let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decode_order(vec![flutter_rust_bridge::for_generated::LockableOrderInfo::new(&api_that, 0, false)]);
@@ -1497,7 +1502,7 @@ let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decod
             }
         }
         let api_that_guard = api_that_guard.unwrap();
- let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::mute_audio(&*api_that_guard))?;   Ok(output_ok)
+ let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::mute_audio(&*api_that_guard, api_source_kind))?;   Ok(output_ok)
                 })()) })
 }
 fn wire__crate__api__dart__api__room__RoomHandle_mute_video_impl(
@@ -1671,7 +1676,8 @@ fn wire__crate__api__dart__api__room__RoomHandle_unmute_audio_impl(
     FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec,_>(flutter_rust_bridge::for_generated::TaskInfo{ debug_name: "RoomHandle_unmute_audio", port: None, mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync }, move || { 
             let message = unsafe { flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(ptr_, rust_vec_len_, data_len_) };
             let mut deserializer = flutter_rust_bridge::for_generated::SseDeserializer::new(message);
-            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);deserializer.end();
+            let api_that = <RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<RoomHandle>>>::sse_decode(&mut deserializer);
+let api_source_kind = <Option<crate::media::track::MediaSourceKind>>::sse_decode(&mut deserializer);deserializer.end();
                 transform_result_sse::<_, ()>((move || {
                     let mut api_that_guard = None;
 let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decode_order(vec![flutter_rust_bridge::for_generated::LockableOrderInfo::new(&api_that, 0, false)]);
@@ -1682,7 +1688,7 @@ let decode_indices_ = flutter_rust_bridge::for_generated::lockable_compute_decod
             }
         }
         let api_that_guard = api_that_guard.unwrap();
- let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::unmute_audio(&*api_that_guard))?;   Ok(output_ok)
+ let output_ok = Result::<_,()>::Ok(crate::api::dart::api::room::RoomHandle::unmute_audio(&*api_that_guard, api_source_kind))?;   Ok(output_ok)
                 })()) })
 }
 fn wire__crate__api__dart__api__room__RoomHandle_unmute_video_impl(