+++ src/room.rs
@@ -513,12 +513,13 @@ impl RoomHandleImpl {
     /// server didn't approve this state transition.
     pub fn mute_audio(
         &self,
+        source_kind: Option<MediaSourceKind>,
     ) -> impl Future<Output = ChangeMediaStateResult> + 'static + use<> {
         self.change_media_state(
             mute_state::Stable::Muted,
             MediaKind::Audio,
             TrackDirection::Send,
-            None,
+            source_kind,
         )
         .map_err(tracerr::map_from_and_wrap!())
     }
@@ -535,12 +536,13 @@ impl RoomHandleImpl {
     /// server didn't approve this state transition.
     pub fn unmute_audio(
         &self,
+        source_kind: Option<MediaSourceKind>,
     ) -> impl Future<Output = ChangeMediaStateResult> + 'static + use<> {
         self.change_media_state(
             mute_state::Stable::Unmuted,
             MediaKind::Audio,
             TrackDirection::Send,
-            None,
+            source_kind,
         )
         .map_err(tracerr::map_from_and_wrap!())
     }
@@ -606,12 +608,13 @@ impl RoomHandleImpl {
     /// server didn't approve this state transition.
     pub fn disable_audio(
         &self,
+        source_kind: Option<MediaSourceKind>,
     ) -> impl Future<Output = ChangeMediaStateResult> + 'static + use<> {
         self.change_media_state(
             media_exchange_state::Stable::Disabled,
             MediaKind::Audio,
             TrackDirection::Send,
-            None,
+            source_kind,
         )
         .map_err(tracerr::map_from_and_wrap!())
     }
@@ -631,12 +634,13 @@ impl RoomHandleImpl {
     /// acquisition request failed.
     pub fn enable_audio(
         &self,
+        source_kind: Option<MediaSourceKind>,
     ) -> impl Future<Output = ChangeMediaStateResult> + 'static + use<> {
         self.change_media_state(
             media_exchange_state::Stable::Enabled,
             MediaKind::Audio,
             TrackDirection::Send,
-            None,
+            source_kind,
         )
         .map_err(tracerr::map_from_and_wrap!())
     }
@@ -709,12 +713,13 @@ impl RoomHandleImpl {
     /// a media server didn't approve this state transition.
     pub fn disable_remote_audio(
         &self,
+        source_kind: Option<MediaSourceKind>,
     ) -> impl Future<Output = ChangeMediaStateResult> + 'static + use<> {
         self.change_media_state(
             media_exchange_state::Stable::Disabled,
             MediaKind::Audio,
             TrackDirection::Recv,
-            None,
+            source_kind,
         )
         .map_err(tracerr::map_from_and_wrap!())
     }
@@ -756,12 +761,13 @@ impl RoomHandleImpl {
     /// a media server didn't approve this state transition.
     pub fn enable_remote_audio(
         &self,
+        source_kind: Option<MediaSourceKind>,
     ) -> impl Future<Output = ChangeMediaStateResult> + 'static + use<> {
         self.change_media_state(
             media_exchange_state::Stable::Enabled,
             MediaKind::Audio,
             TrackDirection::Recv,
-            None,
+            source_kind,
         )
         .map_err(tracerr::map_from_and_wrap!())
     }